+++ src/etools/applications/partners/tests/test_views.py
@@ -21,7 +21,7 @@ from unicef_snapshot.models import Activity
 
 from etools.applications.action_points.models import ActionPoint
 from etools.applications.action_points.tests.factories import ActionPointFactory
-from etools.applications.attachments.tests.factories import AttachmentFileTypeFactory
+from etools.applications.attachments.tests.factories import AttachmentFactory, AttachmentFileTypeFactory
 from etools.applications.core.tests.cases import BaseTenantTestCase
 from etools.applications.core.tests.mixins import URLAssertionMixin
 from etools.applications.funds.models import FundsCommitmentHeader, FundsCommitmentItem
@@ -1663,6 +1663,37 @@ class TestInterventionViews(BaseTenantTestCase):
         self.assertEqual(len(response.data), 1)
         self.assertEqual(response.data[0]["id"], self.intervention["id"])
 
+    def test_intervention_amendment_notificaton(self):
+        attachment = AttachmentFactory()
+        country_programme = CountryProgrammeFactory()
+        office = OfficeFactory()
+        section = SectionFactory()
+        agreement = AgreementFactory(country_programme=country_programme)
+
+        self.active_intervention.in_amendment = True
+        self.active_intervention.agreement = agreement
+        self.active_intervention.country_programme = country_programme
+        self.active_intervention.start = (timezone.now().date()).isoformat()
+        self.active_intervention.unicef_focal_points.add(self.unicef_staff)
+        self.active_intervention.partner_focal_points.add(PartnerStaffFactory())
+        self.active_intervention.save()
+        self.assertEqual(self.active_intervention.status, Intervention.ACTIVE)
+
+        # user is UNICEF User
+        status_code, response = self.run_request(self.active_intervention.id, user=self.partnership_manager_user)
+        self.assertEqual(status_code, status.HTTP_200_OK)
+        self.assertEqual(self.active_intervention.in_amendment, True)
+
+        data = {'in_amendment': False, 'frs': [self.fr_1.id], 'offices': [office.pk], 'sections': [section.pk],
+                'signed_pd_attachment': attachment.pk, 'country_programme': country_programme.pk}
+        status_code, response = self.run_request(
+            self.active_intervention.id,
+            data,
+            user=self.partnership_manager_user,
+            method='patch'
+        )
+        self.active_intervention.refresh_from_db()
+        self.assertEqual(self.active_intervention.in_amendment, False)
 
 class TestInterventionReportingPeriodViews(BaseTenantTestCase):
